refactor: LPページをセクションごとにコンポーネント分割し、Copilotレビュー指摘を対応する
- app/lp/_components/ にセクション単位でコンポーネントを分割
  (HeroSection / DemoSection / HowToSection / PrivacySection / UseCasesSection / FinalCtaSection)
- Before/After画像にsizes属性を追加しモバイルでの過大な画像読み込みを抑止
- ShieldCheckアイコンをemerald-400（緑）に変更（Hero・FinalCTAの2箇所）
- Headerテストにロゴの/リンク遷移テストを追加
- app/lp/page.test.tsxを新規作成（全セクション表示・CTAリンク検証）
